move old constants utility under utils folder
